Fix type error

+++ cpp/src/groupby/hash/compute_global_memory_aggs.cuh
@@ -66,7 +66,7 @@ void compute_global_memory_aggs(cudf::size_type num_rows,
   thrust::for_each_n(
     rmm::exec_policy_nosync(stream),
     thrust::make_counting_iterator(int64_t{0}),
-    num_rows * flattened_values.num_columns(),
+    num_rows * static_cast<int64_t>(flattened_values.num_columns()),
     hash::compute_single_pass_aggs_fn{
       key_indices, *d_values, *d_sparse_table, d_agg_kinds, row_bitmask, skip_rows_with_nulls});
